+++ src/app/components/compare_widget/compare_widget_button.directive.spec.js
@@ -7,7 +7,7 @@
         beforeEach(function () {
             module('chpl.templates');
             module('chpl.compare-widget');
-            inject(function (_$compile_, _$rootScope_, _$log_) {
+            inject(function (_$compile_, _$log_, _$rootScope_) {
                 $compile = _$compile_;
                 $rootScope = _$rootScope_;
                 $log = _$log_;