+++ src/app/components/smart_table/stListMultiple.js
@@ -118,8 +118,9 @@
 
         function findItemWithValue (collection, value) {
             for (var i = 0; i < collection.length; i++) {
-                if (collection[i].value === value)
+                if (collection[i].value === value) {
                     return true;
+                }
             }
             return false;
         }
@@ -155,7 +156,7 @@
     }
 
     /** @ngInclude */
-    function ListMultipleController ($analytics, $log, $localStorage) {
+    function ListMultipleController ($analytics, $localStorage) {
         var vm = this;
 
         vm.activate = activate;